Add files via upload
This is a markdown file that contains the code and explanations of my final project. The question that I tried to answer in my project is which classical music composer has the most articles written about them in the last 50 years in the New York Times. For this project, I have used Python and the New York Times API.